+++ CHANGELOG.md
@@ -4,6 +4,11 @@ To see version information, `xsvutils --version`.
 
 To see compatibility information, `xsvutils help compatibility`.
 
+## version 0.3.3 (2018/05/27)
+- add `rmnoname` subcommand
+- add `insunixtime` subcommand
+- add `--last` option of `cols` subcommand
+
 ## version 0.3.2 (2018/05/21)
 - add `mergecols` subcommand